requirements.txt
asgiref==3.5.2
Django==4.1.1
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.13.1
gunicorn==20.1.0
install==1.3.5
Markdown==3.4.1
Pillow==9.2.0
psycopg2-binary==2.9.3
pytz==2022.2.1
sqlparse==0.4.2
